o  patched endpoints to properly handle millisecond time    stamps (and microsecond).  Casting from double to long    instead of double to int o  removed some references to :: style subsetting in xts,    in favor of only / ISO style.  :: will continue to work    but from a docs perspective will no longer be advertised
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/xts/pkg/xts@603 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/endpoints.c
@@ -47,8 +47,8 @@ SEXP endpoints (SEXP _x, SEXP _on, SEXP _k)
       real_index = REAL(_x);
       ep[0] = 0;
       for(i=1,j=1; i<nr; i++) {
-        real_tmp[0] = (int)real_index[i] / on / k +1;
-        real_tmp[1] = (int)real_index[i-1] / on / k +1;
+        real_tmp[0] = (long)real_index[i] / on / k +1;
+        real_tmp[1] = (long)real_index[i-1] / on / k +1;
         if( (real_tmp[0] - real_tmp[1]) != 0 ) {
           ep[j] = i;
           j++;
